requirements.txt
django==5.2.4
django-filter==25.1